wip: move CommandTranslator into opentrons.file_runner module
