--- api/src/opentrons/protocols/runner/__init__.py
@@ -1,5 +1,0 @@
-from .json_proto.command_translator import CommandTranslator
-
-__all__ = [
-    "CommandTranslator",
-]
